foundationdb_profiling
Struct Mutation

pub struct Mutation {
    pub mutation_type: u8,
    pub param1: Vec<u8>,
    pub param2: Vec<u8>,
}

A mutation (MutationRef), normalized like the C++ deserializer does.

The optional mutation checksum and accumulative checksum index (7.4+, only when the client enables them) are stripped from param2 and their flag bits cleared from mutation_type. A single-key clear, which the wire encodes compactly, is expanded back to ClearRange(key, key + \x00).

Fields
mutation_type: u8

MutationRef::Type code (0 SetValue, 1 ClearRange, 2 AddValue, …).

param1: Vec<u8>

Key, or range begin for ClearRange.

param2: Vec<u8>

Value/operand, or range end for ClearRange.

Implementations

impl Mutation

pub const SET_VALUE: u8 = 0u8

MutationRef::Type::SetValue.

pub const CLEAR_RANGE: u8 = 1u8

MutationRef::Type::ClearRange.

pub const ADD_VALUE: u8 = 2u8

MutationRef::Type::AddValue.

pub const AND: u8 = 6u8

MutationRef::Type::And.

pub const OR: u8 = 7u8

MutationRef::Type::Or.

pub const XOR: u8 = 8u8

MutationRef::Type::Xor.

pub const APPEND_IF_FITS: u8 = 9u8

MutationRef::Type::AppendIfFits.

pub const MAX: u8 = 12u8

MutationRef::Type::Max.

pub const MIN: u8 = 13u8

MutationRef::Type::Min.

pub const SET_VERSIONSTAMPED_KEY: u8 = 14u8

MutationRef::Type::SetVersionstampedKey.

pub const SET_VERSIONSTAMPED_VALUE: u8 = 15u8

MutationRef::Type::SetVersionstampedValue.

pub const BYTE_MIN: u8 = 16u8

MutationRef::Type::ByteMin.

pub const BYTE_MAX: u8 = 17u8

MutationRef::Type::ByteMax.

pub const MIN_V2: u8 = 18u8

MutationRef::Type::MinV2.

pub const AND_V2: u8 = 19u8

MutationRef::Type::AndV2.

pub const COMPARE_AND_CLEAR: u8 = 20u8

MutationRef::Type::CompareAndClear.