Crate foundationdb_profiling

FoundationDB natively implements transaction profiling and analyzing.

The transactions are sampled at the specified rate and all the events for that sampled transaction are recorded. Then at 30 second interval, the data for all the sampled transactions during that interval is flushed to the database. The sampled data is written into special key space \xff\x02/fdbClientInfo/ - \xff\x02/fdbClientInfo0

source

Each data are recorded as chunked events. Events are referenced by the tuple (VersionStamp, TransactionId), then a header defined how much chunks defined the data block.

Profiling Keys look like this:

    FF               - 2 bytes \xff\x02
    SSSSSSSSSS       - 10 bytes Version Stamp
    RRRRRRRRRRRRRRRR - 16 bytes Transaction id
    NNNN             - 4 Bytes Chunk number
    TTTT             - 4 Bytes Total number of chunks

source

To get the data block, each chunk which composes the data block must be accumulated until reaching the “chunk number”

Noted: It could have more than one event in the data block

Constants

PROFILE_PREFIX

Defined by the following pattern

Functions

get_raw_datablocks

Retrieves raw profiling datablocks stored in the database within the specified read version range.